requirements.txt
pyserial>=3.5
numpy>=1.19.0
flask>=2.0.0 